+++ SuperDancer/Classes/Chat/Controller/好友/FriendChatViewController.m
@@ -36,13 +36,17 @@ _Pragma("clang diagnostic pop") \
     [super viewDidAppear:animated];
     if (![[NIMSDK sharedSDK].userManager isMyFriend:_user.userId]) {
         if (!_isAlerted) {
-            [DJWYAlertView showTwoActionAlertViewWithTitle:@"温馨提示" message:@"好友关系已经解除，是否保留会话？" leftButtonTitle:@"删除" rightButtonTitle:@"保留" leftClick:^{
+            UIAlertController *alertContrller = [UIAlertController alertControllerWithTitle:@"温馨提示" message:@"好友关系已经解除，是否保留会话？" preferredStyle:(UIAlertControllerStyleAlert)];
+            
+            UIAlertAction *confirmAction = [UIAlertAction actionWithTitle:@"删除" style:(UIAlertActionStyleDefault) handler:^(UIAlertAction * _Nonnull action) {
                 NIMDeleteMessagesOption *option = [[NIMDeleteMessagesOption alloc] init];
                 option.removeSession = YES;
                 [[NIMSDK sharedSDK].conversationManager deleteAllmessagesInSession:[NIMSession session:_user.userId type:NIMSessionTypeP2P] option:option];
-            } rightClick:^{
-                
             }];
+            UIAlertAction *deleteAction = [UIAlertAction actionWithTitle:@"保留" style:UIAlertActionStyleDefault handler:nil];
+            [alertContrller addAction:deleteAction];
+            [alertContrller addAction:confirmAction];
+            [self presentViewController:alertContrller animated:YES completion:nil];
         }
         
     }
